Converted `expanding_rows.js` to `*.tsx`
