--- src-docs/src/views/tables/expanding_rows/expanding_rows.js
@@ -1,221 +1,0 @@
-import React, { useState, Fragment } from 'react';
-import { formatDate } from '../../../../../src/services/format';
-import { createDataStore } from '../data_store';
-
-import {
-  EuiBasicTable,
-  EuiButtonIcon,
-  EuiHealth,
-  EuiButton,
-  EuiDescriptionList,
-  EuiScreenReaderOnly,
-} from '../../../../../src/components';
-
-import { RIGHT_ALIGNMENT } from '../../../../../src/services';
-
-/*
-Example user object:
-
-{
-  id: '1',
-  firstName: 'john',
-  lastName: 'doe',
-  github: 'johndoe',
-  dateOfBirth: Date.now(),
-  nationality: 'NL',
-  online: true
-}
-
-Example country object:
-
-{
-  code: 'NL',
-  name: 'Netherlands',
-  flag: '🇳🇱'
-}
-*/
-
-const store = createDataStore();
-
-export const Table = () => {
-  const [pageIndex, setPageIndex] = useState(0);
-  const [pageSize, setPageSize] = useState(5);
-  const [sortField, setSortField] = useState('firstName');
-  const [sortDirection, setSortDirection] = useState('asc');
-  const [selectedItems, setSelectedItems] = useState([]);
-  const [itemIdToExpandedRowMap, setItemIdToExpandedRowMap] = useState({});
-
-  const onTableChange = ({ page = {}, sort = {} }) => {
-    const { index: pageIndex, size: pageSize } = page;
-
-    const { field: sortField, direction: sortDirection } = sort;
-
-    setPageIndex(pageIndex);
-    setPageSize(pageSize);
-    setSortField(sortField);
-    setSortDirection(sortDirection);
-  };
-
-  const onSelectionChange = (selectedItems) => {
-    setSelectedItems(selectedItems);
-  };
-
-  const onClickDelete = () => {
-    store.deleteUsers(...selectedItems.map((user) => user.id));
-
-    setSelectedItems([]);
-  };
-
-  const renderDeleteButton = () => {
-    if (selectedItems.length === 0) {
-      return;
-    }
-    return (
-      <EuiButton color="danger" iconType="trash" onClick={onClickDelete}>
-        Delete {selectedItems.length} Users
-      </EuiButton>
-    );
-  };
-
-  const toggleDetails = (item) => {
-    const itemIdToExpandedRowMapValues = { ...itemIdToExpandedRowMap };
-    if (itemIdToExpandedRowMapValues[item.id]) {
-      delete itemIdToExpandedRowMapValues[item.id];
-    } else {
-      const { nationality, online } = item;
-      const country = store.getCountry(nationality);
-      const color = online ? 'success' : 'danger';
-      const label = online ? 'Online' : 'Offline';
-      const listItems = [
-        {
-          title: 'Nationality',
-          description: `${country.flag} ${country.name}`,
-        },
-        {
-          title: 'Online',
-          description: <EuiHealth color={color}>{label}</EuiHealth>,
-        },
-      ];
-      itemIdToExpandedRowMapValues[item.id] = (
-        <EuiDescriptionList listItems={listItems} />
-      );
-    }
-    setItemIdToExpandedRowMap(itemIdToExpandedRowMapValues);
-  };
-
-  const { pageOfItems, totalItemCount } = store.findUsers(
-    pageIndex,
-    pageSize,
-    sortField,
-    sortDirection
-  );
-
-  const deleteButton = renderDeleteButton();
-
-  const columns = [
-    {
-      field: 'firstName',
-      name: 'First Name',
-      sortable: true,
-      truncateText: true,
-      mobileOptions: {
-        render: (item) => (
-          <span>
-            {item.firstName} {item.lastName}
-          </span>
-        ),
-        header: false,
-        truncateText: false,
-        enlarge: true,
-        width: '100%',
-      },
-    },
-    {
-      field: 'lastName',
-      name: 'Last Name',
-      truncateText: true,
-      mobileOptions: {
-        show: false,
-      },
-    },
-    {
-      field: 'dateOfBirth',
-      name: 'Date of Birth',
-      schema: 'date',
-      render: (date) => formatDate(date, 'dobLong'),
-      sortable: true,
-    },
-    {
-      name: 'Actions',
-      actions: [
-        {
-          name: 'Clone',
-          description: 'Clone this person',
-          type: 'icon',
-          icon: 'copy',
-          onClick: () => '',
-        },
-      ],
-    },
-    {
-      align: RIGHT_ALIGNMENT,
-      width: '40px',
-      isExpander: true,
-      name: (
-        <EuiScreenReaderOnly>
-          <span>Expand rows</span>
-        </EuiScreenReaderOnly>
-      ),
-      render: (item) => (
-        <EuiButtonIcon
-          onClick={() => toggleDetails(item)}
-          aria-label={itemIdToExpandedRowMap[item.id] ? 'Collapse' : 'Expand'}
-          iconType={
-            itemIdToExpandedRowMap[item.id] ? 'arrowDown' : 'arrowRight'
-          }
-        />
-      ),
-    },
-  ];
-
-  const pagination = {
-    pageIndex: pageIndex,
-    pageSize: pageSize,
-    totalItemCount: totalItemCount,
-    pageSizeOptions: [3, 5, 8],
-  };
-
-  const sorting = {
-    sort: {
-      field: sortField,
-      direction: sortDirection,
-    },
-  };
-
-  const selection = {
-    selectable: (user) => user.online,
-    selectableMessage: (selectable) =>
-      !selectable ? 'User is currently offline' : undefined,
-    onSelectionChange: onSelectionChange,
-  };
-
-  return (
-    <Fragment>
-      {deleteButton}
-      <EuiBasicTable
-        tableCaption="Demo of EuiBasicTable with expanding rows"
-        items={pageOfItems}
-        itemId="id"
-        itemIdToExpandedRowMap={itemIdToExpandedRowMap}
-        isExpandable={true}
-        hasActions={true}
-        columns={columns}
-        pagination={pagination}
-        sorting={sorting}
-        isSelectable={true}
-        selection={selection}
-        onChange={onTableChange}
-      />
-    </Fragment>
-  );
-};
